requirements.txt
python-jose[cryptography]
fastapi[standard]
passlib[bcrypt]
flask
flask-sqlalchemy
python-multipart